# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/roles": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取所有角色的列表",
                "tags": [
                    "RoleApi"
                ],
                "summary": "获取角色列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "添加一个新的角色",
                "tags": [
                    "RoleApi"
                ],
                "summary": "添加角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "角色状态",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "角色代码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "排序值",
                        "name": "sort",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据角色ID获取角色的详细信息",
                "tags": [
                    "RoleApi"
                ],
                "summary": "获取角色详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据角色ID编辑角色信息",
                "tags": [
                    "RoleApi"
                ],
                "summary": "编辑角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "角色名称",
                        "name": "RoleName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "角色密码",
                        "name": "Password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据角色ID删除角色",
                "tags": [
                    "RoleApi"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应编码：0成功 401请登录 403无权限 500错误",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据对象"
                },
                "msg": {
                    "description": "消息提示",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}